arm64: dts: freescale: Add support for the headless GOcontroll Moduline controllers
The GOcontroll Moduline IV and Mini are embedded controllers that allow
for reconfigurable in/outputs via IO modules. Further expandable with
wifi/bluetooth and LTE/GPS. The default interfaces consist of ethernet
and a usb port which can also be configured as a RS232 interface.

The Moduline IV:
 * 8 IO module slots
 * 4 CAN busses
 * 1 LIN bus

The Moduline Mini:
 * 4 IO module slots
 * 2 CAN busses
 * 3 axis accelerometer

Signed-off-by: Maud Spierings <maudspierings@gocontroll.com>