dt-bindings: connector: Add the GOcontroll Moduline module slot bindings
Add the bindings that describe a GOcontroll Moduline module slot. This
slot provides all the interfaces to interface with a Moduline compatible
IO module. The actual module is not reasonable to describe as it can be
swapped at will, with this connecter the driver will be able to probe
for a module on boot.

Signed-off-by: Maud Spierings <maudspierings@gocontroll.com>